Added first version of avrora.sim.state.Register, added new code from Pekka Nikander

+++ src/avrora/sim/mcu/ATMega128New.java
@@ -426,7 +426,7 @@ public class ATMega128New extends ATMegaClassic {
     private static final int[] MCUCR_sm_perm = { 2, 4, 3 };
 
     protected int getSleepMode() {
-	byte value = MCUCR_reg.read();
+	byte value = (byte)MCUCR_reg.read();
         boolean sleepEnable = Arithmetic.getBit(value, 5);
 
         if ( sleepEnable )
